pyproject.toml
[tool.poetry]
name = "n5now"
version = "0.1.0"
description = ""
authors = ["Facundo Padilla <facundo.padilla@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
pydantic-partial = "^0.5.5"
fastapi-pagination = "^0.12.26"
orjson = "^3.10.6"


[tool.poetry.group.testing.dependencies]
pytest = "^8.2.2"
respx = "^0.21.1"
pytest-asyncio = "^0.23.7"
aiosqlite = "^0.20.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.database.dependencies]
sqlmodel = "^0.0.19"
alembic = "^1.13.2"
asyncpg = "^0.29.0"


[tool.poetry.group.security.dependencies]
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^42.0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
